生码赋值新增产品名称；退货详情；查询流转节点对应码明细

+++ acc-module/src/main/java/com/ztl/gym/storage/service/IStorageInService.java
@@ -111,5 +111,5 @@ public interface IStorageInService {
      * @param storageBack
      * @return
      */
-    int insertStorageInForBack(StorageBack storageBack);
+    long insertStorageInForBack(StorageBack storageBack);
 }
